small fixes and comments

+++ Program.cs
@@ -9,11 +9,13 @@ using WixSharp;
 using WixSharp.CommonTasks;
 
 await Parser.Default.ParseArguments<Options>(args)
-    .WithNotParsed(_ =>
+    .WithNotParsed(errors =>
     {
-        Console.WriteLine("Error: failed to parse command line arguments");
-
-        Environment.Exit(-1);
+        if (errors.Any())
+        {
+            Console.WriteLine("Error: invalid arguments");
+            Environment.Exit(-1);
+        }
     })
     .WithParsedAsync(Run);
 
@@ -26,12 +28,14 @@ static async Task Run(Options options)
         Console.WriteLine();
     }
 
+    // check if archive is specified
     if (string.IsNullOrWhiteSpace(options.ArchivePath))
     {
         options.IfVerbose(() => Console.WriteLine($"No archive specified, using \"{Constants.DefaultVSCodeArchiveUrl}\""));
         options.ArchivePath = Constants.DefaultVSCodeArchiveUrl;
     }
 
+    // check if archive is a URL, and if so, download it
     options.IfVerbose(() => Console.WriteLine("Checking if archive should be downloaded..."));
     if (Uri.TryCreate(options.ArchivePath, UriKind.Absolute, out var uri))
     {
@@ -53,6 +57,7 @@ static async Task Run(Options options)
         };
         webClient.DownloadFileCompleted += (_, e) => { lock (locker) { Console.WriteLine(" Done."); } };
 
+        // download file
         await webClient.DownloadFileTaskAsync(uri, filename);
 
         options.ArchivePath = filename;
@@ -66,6 +71,7 @@ static async Task Run(Options options)
     }
 
     options.IfVerbose(() => Console.WriteLine("Checking for existing extracted archive..."));
+    // if archive was extracted before, remove old extracted archive
     if (Directory.Exists(Constants.ArchiveExtractPath))
     {
         Console.WriteLine("Removing old extracted archive...");
@@ -77,9 +83,14 @@ static async Task Run(Options options)
     Console.WriteLine(" Done.");
     options.IfVerbose(() => Console.WriteLine($"Archive extracted to \"{Constants.ArchiveExtractPath}\""));
 
-    options.IfVerbose(() => Console.WriteLine("Removing unneeded archive..."));
-    System.IO.File.Delete(options.ArchivePath);
+    // remove unneeded archive if it was not specified
+    if (options.ArchivePath == Constants.ArchiveDownloadPath)
+    {
+        options.IfVerbose(() => Console.WriteLine("Removing unneeded archive..."));
+        System.IO.File.Delete(options.ArchivePath);
+    }
 
+    // locating vscode executable
     var codeExe = Path.Combine(Constants.ArchiveExtractPath, "Code.exe");
     options.IfVerbose(() => Console.WriteLine("Checking for vscode executable..."));
     if (!System.IO.File.Exists(codeExe))
@@ -89,6 +100,7 @@ static async Task Run(Options options)
     }
     options.IfVerbose(() => Console.WriteLine($"VSCode executable is at \"{codeExe}\""));
 
+    // extract icon of vscode executable
     options.IfVerbose(() => Console.WriteLine("Extracting icon from vscode executable..."));
     var iconFile = ExtractIconFile(codeExe);
     if (iconFile is null)
@@ -97,9 +109,11 @@ static async Task Run(Options options)
     }
     options.IfVerbose(() => Console.WriteLine($"Successfully extracted icon to \"{iconFile}\""));
 
+    // get vscode attributes (version, name, etc.)
     options.IfVerbose(() => Console.WriteLine("Extracting attributes from vscode executable..."));
     var attributes = FileVersionInfo.GetVersionInfo(codeExe);
 
+    // computing correct output path as expected by WiX
     if (string.IsNullOrWhiteSpace(options.OutputPath))
     {
         options.OutputPath = Environment.CurrentDirectory + Path.DirectorySeparatorChar + $"VSCode-{attributes.ProductVersion}-x64";
@@ -116,25 +130,31 @@ static async Task Run(Options options)
     }
 
     options.IfVerbose(() => Console.WriteLine("Configuring msi..."));
+    // base structure
     Project msi = new(attributes.ProductName,
         new Dir(@"%ProgramFiles%\Microsoft VS Code",
                 new Files(Constants.ArchiveExtractPath + Path.DirectorySeparatorChar + "*.*")));
 
     msi.ResolveWildCards();
+    // add desktop and app menu shortcuts
     msi.FindFile(f => f.Name.EndsWith("Code.exe"))[0]
         .AddShortcuts(
             new FileShortcut(attributes.ProductName, "ProgramMenuFolder"),
             new FileShortcut(attributes.ProductName, "%Desktop%"));
     msi.MajorUpgradeStrategy = MajorUpgradeStrategy.Default;
+    // add vscode binaries to PATH
     msi.Add(new EnvironmentVariable("PATH", @"[INSTALLDIR]\bin") { Part = EnvVarPart.last });
 
+    // basic attributes
     msi.Version = new(attributes.ProductMajorPart, attributes.ProductMinorPart, attributes.ProductBuildPart);
     msi.Platform = Platform.x64;
     msi.OutFileName = options.OutputPath;
     msi.LicenceFile = Constants.ArchiveExtractPath + Path.DirectorySeparatorChar + @"resources\app\LICENSE.rtf";
 
+    // set GUID of installer
     msi.GUID = new Guid("fcd5a47f-9d70-4c32-8c3a-ae65c9b17a64");
 
+    // advanced attributes
     msi.ControlPanelInfo.NoModify = true;
     msi.ControlPanelInfo.Comments = attributes.FileDescription;
     msi.ControlPanelInfo.Readme = "https://code.visualstudio.com/learn";
@@ -143,6 +163,7 @@ static async Task Run(Options options)
     msi.ControlPanelInfo.Manufacturer = attributes.CompanyName;
     msi.ControlPanelInfo.InstallLocation = "[INSTALLDIR]";
 
+    // if icon file was extracted use it as product icon
     if (iconFile is not null)
     {
         msi.ControlPanelInfo.ProductIcon = iconFile;
@@ -150,13 +171,14 @@ static async Task Run(Options options)
 
     Console.Write("Building msi...");
     Compiler.BuildMsi(msi);
-    Console.WriteLine(" Done.");
+    Console.WriteLine("Successfully built MSI installer.");
 }
 
 static string? ExtractIconFile(string path)
 {
     try
     {
+        // we try to use IconLib first
         MultiIcon multiIcon = [];
         multiIcon.Load(path);
         multiIcon.Save(Constants.VSCodeIconPath, MultiIconFormat.ICO);
@@ -164,6 +186,7 @@ static string? ExtractIconFile(string path)
     catch
     {
         Console.WriteLine("Warning: failed to extract icon via IconLib, falling back to System.Drawing");
+        // if this fails we try to use System.Drawing, this icon won't be as good but it's better than nothing
         using var icon = Icon.ExtractAssociatedIcon(path);
 
         if (icon is null)
